Extent Reports populated now. Also, the test cases grouped to BeforeMidNight and NextDay.
