--- DSARform_ManualUpload_DataDriven.csv
@@ -1,3 +1,0 @@
-﻿Mr,Hasan,Gholi,Kachal,10,10,1990
-Mrs,Roobahe,Gholi,Makaar,11,11,1991
-Miss,Gorbe,Gholi,Nare,12,12,1992
