+++ NextDay.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿//using Microsoft.VisualStudio.TestTools.UnitTesting;
+using NUnit.Framework;
 using OpenQA.Selenium;
 using OpenQA.Selenium.Appium.Windows;
 using OpenQA.Selenium.Chrome;
@@ -25,48 +26,34 @@ using WindowsInput.Native;
 
 namespace WebUITestAutomation
 {
-    [TestClass]
-    public class LibertyJobSearch
+    
+    
+    
+    
+    [TestFixture]
+    public class NextDay : Base
     {
-        float pageCountCompleted;
+
+
+
+        float pageCountCompleted_Liberty;
         //string WinDriver = "C:\\Program Files\\Windows Application Driver\\WinAppDriver.exe";
         string BaseDirectory_Path = AppDomain.CurrentDomain.BaseDirectory;
+        String WinDriver = "C:\\Program Files\\Windows Application Driver\\WinAppDriver.exe";
 
 
-        IWebDriver driver;
+        //IWebDriver driver;
         FirstPublicMap Public;
         CurrentOpportunitiesMap Opportunities;
         SearchMap SearchP;
         PersonalInfoMap PersonalInfo;
 
 
-        [TestMethod]
+        [Test]
         public void PendingAppForPortfolioManagerPosition()
         {
-            //Run WinAppDriver
-            // AppDomain.CurrentDomain.BaseDirectory this is here: C:\Users\44741\source\repos\UnitTestProject1\bin\Debug . it is part of the solution folder
-            /*******
-            //Files need to be copied to C:\Users\44741\source\repos\UnitTestProject1 . when the solution is built, a copy of the files from C:\Users\44741\source\repos\UnitTestProject1 will be copied to C:\Users\44741\source\repos\UnitTestProject1\bin\Debug where the file will be read/run automatically.
-            *******/
-            string WinDriver_FullPath = BaseDirectory_Path + "\\WinAppDriver_baseDirectory\\WinAppDriver.exe";
-            Process.Start(@WinDriver_FullPath);
-            //Process.Start(@WinDriver);
-            //Process.Start($"WinAppDriver.exe", AppDomain.CurrentDomain.BaseDirectory); 
-            DesiredCapabilities desktopCapabilities = new DesiredCapabilities();
-            desktopCapabilities.SetCapability("app", "Root");
-            WindowsDriver<WindowsElement> session = new WindowsDriver<WindowsElement>(new Uri("http://127.0.0.1:4723"), desktopCapabilities);
+            
 
-
-            //to kick start Selenium ChromeDriver
-            new WebDriverManager.DriverManager().SetUpDriver(new ChromeConfig(), VersionResolveStrategy.MatchingBrowser);
-            var options = new ChromeOptions();
-            options.AddArgument("no-sandbox");
-
-
-            //starting browser
-            ChromeDriver driver = new ChromeDriver(options);
-            driver.Manage().Window.Maximize();
-            driver.Manage().Timeouts().ImplicitWait = TimeSpan.FromSeconds(20);
             driver.Navigate().GoToUrl("https://www.libertyspecialtymarkets.com/gb-en");
 
             var Cookies = new WebDriverWait(driver, TimeSpan.FromSeconds(30));
@@ -120,18 +107,18 @@ namespace WebUITestAutomation
 
             if (pageCount <= 1)
             {
-                pageCountCompleted = 1;
+                pageCountCompleted_Liberty = 1;
             }
             else
             {
-                pageCountCompleted = pageCount + 1;
+                pageCountCompleted_Liberty = pageCount + 1;
             }
 
 
             bool isFind = false;
             int l = 0;
 
-            for (int k = 0; k < pageCountCompleted; k++)
+            for (int k = 0; k < pageCountCompleted_Liberty; k++)
             {
 
                 IList<IWebElement> positions = driver.FindElements(By.XPath("/html/body/div/main/div/div/section/div[3]/div[2]/article/div/div/h3/a"));
@@ -234,7 +221,7 @@ namespace WebUITestAutomation
 
             //Console.WriteLine("Application succesfully submitted. Good luck!");
 
-            //driver.Quit();
+            driver.Quit();
 
 
 
@@ -311,28 +298,11 @@ namespace WebUITestAutomation
         }
 
 
-        [TestMethod]
+        [Test]
         public void SubmitAppForPortfolioManagerPosition()
         {
-            //Run WinAppDriver
-            // AppDomain.CurrentDomain.BaseDirectory this is here: C:\Users\44741\source\repos\UnitTestProject1\bin\Debug . it is part of the solution folder
-            string WinDriver_FullPath = BaseDirectory_Path + "\\WinAppDriver_baseDirectory\\WinAppDriver.exe";
-            Process.Start(@WinDriver_FullPath);
-            DesiredCapabilities desktopCapabilities= new DesiredCapabilities();
-            desktopCapabilities.SetCapability("app", "Root");
-            WindowsDriver<WindowsElement> session = new WindowsDriver<WindowsElement>(new Uri("http://127.0.0.1:4723"), desktopCapabilities);
+            
 
-
-            //to kick start Selenium ChromeDriver
-            new WebDriverManager.DriverManager().SetUpDriver(new ChromeConfig(), VersionResolveStrategy.MatchingBrowser);
-            var options = new ChromeOptions();
-            options.AddArgument("no-sandbox");
-
-
-            //starting browser
-            ChromeDriver driver = new ChromeDriver(options);
-            driver.Manage().Window.Maximize();
-            driver.Manage().Timeouts().ImplicitWait = TimeSpan.FromSeconds(20);
             driver.Navigate().GoToUrl("https://www.libertyspecialtymarkets.com/gb-en");
 
             var Cookies = new WebDriverWait(driver, TimeSpan.FromSeconds(30));
@@ -386,18 +356,18 @@ namespace WebUITestAutomation
 
             if (pageCount <= 1)
             {
-                pageCountCompleted = 1;
+                pageCountCompleted_Liberty = 1;
             }
             else
             {
-                pageCountCompleted = pageCount + 1;
+                pageCountCompleted_Liberty = pageCount + 1;
             }
 
 
             bool isFind = false;            
             int l = 0;
 
-            for (int k = 0; k < pageCountCompleted; k++)
+            for (int k = 0; k < pageCountCompleted_Liberty; k++)
             {
 
                 IList<IWebElement> positions = driver.FindElements(By.XPath("/html/body/div/main/div/div/section/div[3]/div[2]/article/div/div/h3/a"));
@@ -555,16 +525,288 @@ namespace WebUITestAutomation
             // 5     /html/body/div/main/div/div/section/div[3]/div[1]/div[3]/a[4]
             // 6     /html/body/div/main/div/div/section/div[3]/div[1]/div[3]/a[5]
 
+        }
+
+
+        [Test]
+        public void FindBook12()
+        {
+            float pageCountCompleted;
+
+
+            driver.Navigate().GoToUrl("https://www.bahaibookstore.com/");
+            //wait until DSAR Form ison is visible
+            Thread.Sleep(2000);
+
+
+            // list of products
+            IList<IWebElement> ProductCategories = driver.FindElementsByXPath("/html/body/form/div[3]/div/div[4]/div[1]/div/div/div[1]/div/div[2]/ul/li/a");
+            //click on Free Downloads
+            int i = 1;
+            foreach (var ProductCategorie in ProductCategories)
+            {
+                if (ProductCategorie.Text.Equals("Free Downloads"))
+                {
+                    // hover over Free Downloads
+                    Actions HoverOver = new Actions(driver);
+                    //Performing the mouse hover action on the target element.
+                    HoverOver.MoveToElement(driver.FindElement(By.XPath("/html/body/form/div[3]/div/div[4]/div[1]/div/div/div[1]/div/div[2]/ul/li[" + i + "]/a"))).Perform();
+                    break;
+                }
+                i++;
+            }
+
+            Thread.Sleep(2000);
+
+            // list of Languages under Free Downloads
+            IList<IWebElement> Languages = driver.FindElementsByXPath("/html/body/form/div[3]/div/div[4]/div[1]/div/div/div[1]/div/div[2]/ul/li[13]/ul/li/a");
+            //click on English
+            int j = 1;
+            foreach (var Language in Languages)
+            {
+                if (Language.Text.Equals("English"))
+                {
+                    driver.FindElement(By.XPath("/html/body/form/div[3]/div/div[4]/div[1]/div/div/div[1]/div/div[2]/ul/li[13]/ul/li[" + j + "]/a")).Click();
+                    break;
+                }
+                j++;
+            }
+
+            var English = new WebDriverWait(driver, TimeSpan.FromSeconds(30));
+            English.Until(ExpectedConditions.ElementIsVisible(By.XPath("(/html/body/form/div[3]/div/div[3]/div[2]/div/div[1]/div/div[2]/div/h1)")));
 
 
 
+            string total = driver.FindElementByXPath("/html/body/form/div[3]/div/div[3]/div[2]/div/div[1]/div/div[3]/div[1]/div/table/tbody/tr/td[1]/div[1]/span").Text;
+            string[] count = total.Split(' ');
+            //string[] totalValues = count[0].Split('1');
+            float pageCount = int.Parse(count[5]) / 12;
 
 
+            if (pageCount <= 1)
+            {
+                pageCountCompleted = 1;
+            }
+            else
+            {
+                pageCountCompleted = pageCount + 1;
+            }
 
+            bool isFind = false;
+            //int l = 0;
+            int d = 0;
+
+            for (int k = 0; k < pageCountCompleted; k++)
+            {
+                int l = 0;
+                IList<IWebElement> TitleOfBooks = driver.FindElements(By.XPath("/html/body/form/div[3]/div/div[3]/div[2]/div/div[1]/div/div[3]/div[2]/div[1]/div/table/tbody/tr/td/div/table/tbody/tr/td/div/a/div"));
+
+                foreach (IWebElement TitleOfBook in TitleOfBooks)
+                {
+
+                    l++;
+                    if (TitleOfBook.Text.Contains("Creating a New Mind (Free ePub)"))
+                    {
+                        string isPrice = driver.FindElementByXPath("/html/body/form/div[3]/div/div[3]/div[2]/div/div[1]/div/div[3]/div[2]/div[1]/div/table/tbody/tr[" + l + "]/td/div/table/tbody/tr/td/div/a/div/../../span[8]").Text.Trim();
+                        Assert.AreEqual(isPrice, "$0.00");
+                        isFind = true;
+                        //click on Apply for Portfolio Manager
+                        driver.FindElementByXPath("/html/body/form/div[3]/div/div[3]/div[2]/div/div[1]/div/div[3]/div[2]/div[1]/div/table/tbody/tr[" + l + "]/td/div/table/tbody/tr/td/div/a/div").Click();
+                        break;
+                    }
+                }
+                if (isFind)
+                {
+                    break;
+                }
+                // Go to next page
+                d = k + 2;
+                driver.FindElementByXPath("/html/body/form/div[3]/div/div[3]/div[2]/div/div[1]/div/div[3]/div[1]/div/table/tbody/tr/td[1]/div[2]/a[" + d + "]").Click();
+                Thread.Sleep(3000);
+
+
+            }
+
+            Assert.IsTrue(isFind, "Your desired free downloadable book is not listed. Sorry!");
+
+
+            driver.Quit();
 
 
 
         }
+
+
+
+
+
+        [Test]
+        public void SampleSpareCodes()
+        {
+
+
+            //to run WinAppDriver
+            Process.Start(@WinDriver);
+            DesiredCapabilities desktopCapabilities = new DesiredCapabilities();
+            desktopCapabilities.SetCapability("app", "Root");
+            WindowsDriver<WindowsElement> session = new WindowsDriver<WindowsElement>(new Uri("http://127.0.0.1:4723"), desktopCapabilities);
+
+            //to kick start Selenium ChromeDriver
+            new WebDriverManager.DriverManager().SetUpDriver(new ChromeConfig(), VersionResolveStrategy.MatchingBrowser);
+            var options = new ChromeOptions();
+            options.AddArgument("no-sandbox");
+            //starting browser
+            driver = new ChromeDriver(options);
+            driver.Manage().Window.Maximize();
+            driver.Manage().Timeouts().ImplicitWait = TimeSpan.FromSeconds(10);
+            driver.Navigate().GoToUrl("https://www.kayak.co.uk/flights");
+
+
+            driver.FindElementByXPath("").Click();
+            Thread.Sleep(3000);
+
+            session.FindElementByXPath("").Click();
+
+            Thread.Sleep(2000);
+
+            InputSimulator sim = new InputSimulator();
+            sim.Keyboard.KeyPress(VirtualKeyCode.DOWN);
+            sim.Keyboard.KeyPress(VirtualKeyCode.DOWN);
+            sim.Keyboard.KeyPress(VirtualKeyCode.DOWN);
+            sim.Keyboard.KeyPress(VirtualKeyCode.DOWN);
+            sim.Keyboard.KeyPress(VirtualKeyCode.DOWN);
+            sim.Keyboard.KeyPress(VirtualKeyCode.RETURN);
+
+            Thread.Sleep(2000);
+
+            driver.SwitchTo().Alert().Accept();
+
+            driver.SwitchTo().Frame(0);
+
+
+            //----------------------------------
+
+            string totalCases = driver.FindElement(By.XPath("//legend[contains(text(),'Cases')]")).Text;
+            string[] count = totalCases.Split(' ');
+            string[] totalValues = count[1].Split('(');
+            float pageCount = int.Parse(totalValues[1]) / 10;
+            float pageCountCompleted = pageCount + 1;
+            string caseID = "1110011";
+            List<string> DMR = new List<string>();
+            DMR.Add("C01A"); DMR.Add("C01B"); DMR.Add("C02A"); DMR.Add("C02B");
+            DMR.Add("C02C"); DMR.Add("C02D"); DMR.Add("C02E"); DMR.Add("C02F");
+            DMR.Add("C09B");
+
+            bool isFind = false;
+            int caseNumber = 0;
+
+            for (int i = 1; i <= pageCountCompleted; i++)
+            {
+                IList<IWebElement> allCases = driver.FindElements(By.XPath("//strong//a"));
+                for (int j = 0; j < allCases.Count; j++)
+                {
+                    if (allCases[j].Text.Equals(caseID))
+                    {
+                        isFind = true;
+                        Console.WriteLine("case " + caseID + " found");
+                        caseNumber = j + 1;
+                        break;
+                    }
+                }
+                if (isFind)
+                    break;
+                else
+                {
+                    driver.FindElement(By.XPath("//a[text()='" + (i + 1) + "']")).Click();
+                }
+            }
+
+            Assert.IsTrue(isFind, "Unable to find the CaseID. Please check your CaseID:" + caseID);
+
+            IList<IWebElement> elements = driver.FindElements(By.XPath("//div//strong//a[text()='" + caseID + "']/../../..//span[@title='Reason for filing']//following-sibling::span"));
+            for (int i = 0; i < DMR.Count; i++)
+            {
+                bool isMatch = false;
+                for (int j = 0; j < elements.Count; j++)
+                {
+                    if (elements[j].Text.Contains(DMR[i]))
+                    {
+                        isMatch = true;
+                        break;
+                    }
+                }
+                if (isMatch == false)
+                {
+                    Assert.Fail("DMR value is not matched to the corresponsing case number. Please check the case number is:" + caseID);
+                }
+            }
+
+            //----------------------------------
+
+            //   Assert.IsTrue(findpage.GetDBName().Contains("Internal Fraud Database"), "Unable to change the database. please check.");   
+
+
+            //------------------------------------
+
+
+            string totalFrauds = driver.FindElement(By.XPath("//legend[contains(text(),'Frauds')]")).Text;
+            //   string[] count = totalFrauds.Split(' ');
+            //   string[] totalValues = count[1].Split('(');
+            //   float pageCount = int.Parse(totalValues[1]) / 10;
+            //    float pageCountCompleted = pageCount + 1;
+            string fraudID = "1001";
+
+            //  bool isFind = false;
+
+            for (int i = 1; i <= pageCountCompleted; i++)
+            {
+                IList<IWebElement> allCases = driver.FindElements(By.XPath("//b//a"));
+                for (int j = 0; j < allCases.Count; j++)
+                {
+                    if (allCases[j].Text.Equals(fraudID))
+                    {
+                        isFind = true;
+                        Console.WriteLine("fraud " + fraudID + " found");
+
+                        break;
+                    }
+                }
+                if (isFind)
+                    break;
+                else
+                {
+                    driver.FindElement(By.XPath("//a[text()='" + (i + 1) + "']")).Click();
+                    Thread.Sleep(5000);
+                }
+            }
+
+            Assert.IsTrue(isFind, "Unable to find the fraudID. Please check your FraudID:" + fraudID);
+
+
+            //-------------------------------------
+
+            ChromeOptions chromeOptions = new ChromeOptions();
+            chromeOptions.AddAdditionalCapability("pageLoadStrategy", "none");
+            driver = new ChromeDriver($"{AppDomain.CurrentDomain.BaseDirectory}", chromeOptions);
+            driver = new ChromeDriver($"{AppDomain.CurrentDomain.BaseDirectory}");
+            driver.Manage().Window.Maximize();
+            driver.Navigate().GoToUrl("https://www.google.co.uk");
+
+            //---------------------------------------
+
+            Assert.AreEqual("Mr Search Search - 01 Jan 2000", driver.FindElement(By.XPath("")).Text);
+
+            //----------------------------------------
+
+
+
+        }
+
+
+
+
+
+
     }
 
 }
