+++ packages.config
@@ -74,6 +74,7 @@
   <package id="MSTest.TestFramework" version="2.2.10" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net472" />
   <package id="NUnit" version="3.13.3" targetFramework="net472" />
+  <package id="NUnit3TestAdapter" version="4.4.2" targetFramework="net472" />
   <package id="RazorEngine.NetCore" version="2.2.6" targetFramework="net472" />
   <package id="Selenium.Support" version="3.11.0" targetFramework="net472" />
   <package id="Selenium.WebDriver" version="3.11.0" targetFramework="net472" />
